common SQL to base
